requirements.txt
# # Core Dependencies
# pdfminer.six==20221105
# PyMuPDF==1.23.8
# # fitz==0.0.1.dev2 -- error version

# # ML and Embeddings
# sentence-transformers==2.2.2
# faiss-cpu==1.7.4
# torch>=2.0.0
# transformers>=4.30.0

# # LLM Integration
# openai==1.12.0
# anthropic==0.18.1
# cohere==4.47

# # API Framework
# fastapi==0.109.2
# uvicorn[standard]==0.27.1
# pydantic==2.6.1
# pydantic-settings==2.1.0
# python-multipart==0.0.9

# # Utilities
# python-dotenv==1.0.1
# pyyaml==6.0.1
# numpy==1.24.3
# pandas==2.0.3
# tqdm==4.66.1
# colorlog==6.8.2

# # Testing and Evaluation
# pytest==7.4.4
# pytest-asyncio==0.23.4
# scikit-learn==1.3.2

# # Development
# black==24.1.1
# flake8==7.0.0
# mypy==1.8.0

# Core Dependencies
pdfminer.six==20221105
PyMuPDF==1.23.8

# ML and Embeddings - UPDATED VERSIONS
sentence-transformers>=2.3.0
huggingface-hub>=0.20.0
faiss-cpu==1.7.4
torch>=2.0.0
transformers>=4.30.0

# LLM Integration
openai>=1.12.0
anthropic>=0.18.1
cohere>=4.47

# API Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-multipart>=0.0.9

# Utilities
python-dotenv>=1.0.1
pyyaml>=6.0.1
numpy>=1.24.3,<2.0.0
pandas>=2.0.3
tqdm>=4.66.1

# Testing and Evaluation
pytest>=8.0.0
pytest-asyncio>=0.23.4
scikit-learn>=1.3.2

# Development
black>=24.1.1
flake8>=7.0.0
mypy>=1.8.0

requests>=2.31.0
